# use type parameter for E matrix