If the field `E` is not associated with a type parameter, each access of `E` is type unstable, leading to a lot of inference problems and long compile times. This change brings a simple benchmark I have (using RobustAndOptimalControl which calls `ghinfnorm` in this package`) from 
```julia
julia> using RobustAndOptimalControl, ControlSystems

julia> G = ssrand(2,2,2);

julia> @time hinfnorm2(G);
 18.141708 seconds (50.01 M allocations: 2.384 GiB, 8.45% gc time, 99.97% compilation time)
```
to
```
julia> @time hinfnorm2(G);
  6.451567 seconds (26.34 M allocations: 1.337 GiB, 7.95% gc time, 99.99% compilation time)
```

This is a step towards addressing the latest findings from 
- https://github.com/andreasvarga/DescriptorSystems.jl/issues/9