@andreasvarga 2022-03-09
This seems to be an useful feature. Have you tried jointly with
`Base.hcat(systems::DescriptorStateSpace...) ` ?