@baggepinnen 2022-03-09
I haven't tried the suggested change, there were a lot of places that needed similar updates so I wanted to discuss the strategy before proceeding. I'll have a go at the suggested strategy tomorrow