@baggepinnen 2022-03-10
With the latest commit, tests pass locally, and this simple compile-time benchmark goes from
```julia
julia> using DescriptorSystems;

julia> G = rss(2,2,4);

julia> @time ghinfnorm(G);
 20.182468 seconds (59.23 M allocations: 2.710 GiB, 5.65% gc time, 100.00% compilation time)
```
to
```julia
julia> @time ghinfnorm(G);
  6.971881 seconds (24.47 M allocations: 1.197 GiB, 10.94% gc time, 99.99% compilation time)
```

To verify that we didn't get any run-time regressions, here's a benchmark of the runtime as well
Before:
```julia
julia> using BenchmarkTools

julia> @btime ghinfnorm($G);
  1.535 μs (23 allocations: 2.11 KiB)
```

After:
```julia
julia> @btime ghinfnorm($G);
  1.419 μs (23 allocations: 2.20 KiB)
```

All tests were performed on julia Version 1.8.0-beta1